Reorganise the test suite fixture layout since it's starting to get bloated
