+++ tests/fixtures/conftest_group_layers.py
@@ -3,7 +3,6 @@ from typing import Any
 
 import pytest
 from napari_experimental.group_layer import GroupLayer
-from napari_experimental.group_layer_qt import QtGroupLayerModel
 
 
 def return_copy_with_new_name(obj: Any, new_name: str = "Copy") -> Any:
@@ -56,13 +55,3 @@ def nested_group_data(points_layer) -> GroupLayer:
 
     root = GroupLayer(points_0, group_a, points_1, group_b)
     return root
-
-
-def test_qt_group_layer_model(
-    group_layer_data: GroupLayer, nested_group_data: GroupLayer, qtmodeltester
-) -> None:
-    simple = QtGroupLayerModel(root=group_layer_data)
-    qtmodeltester.check(simple)
-
-    nested = QtGroupLayerModel(nested_group_data)
-    qtmodeltester.check(nested)
